telegram-bot/requirements.txt
python-telegram-bot
google-api-core
google-cloud-compute
googleapis-common-protos